pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sandbox-runtime"
version = "0.1.0"
description = "A lightweight sandboxing tool for enforcing filesystem and network restrictions on arbitrary processes at the OS level"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Anthropic" }
]
keywords = ["sandbox", "security", "process-isolation", "macos", "linux"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Operating System Kernels :: Linux",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
]

[project.scripts]
srt-py = "sandbox_runtime.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "pytest-timeout>=2.0",
    "mypy>=1.0",
    "ruff>=0.1",
    "pyright>=1.1",
    "pre-commit>=3.0",
    "aioresponses>=0.7",
    "httpx>=0.25",
]

[tool.hatch.build.targets.wheel]
packages = ["sandbox_runtime"]
include = [
    "vendor/",
]

[tool.hatch.build.targets.sdist]
include = [
    "sandbox_runtime/",
    "vendor/",
]

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = ["README.md"]

[tool.ruff.lint]
select = [
    "C4",   # flake8-comprehensions
    "C90",  # mccabe
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "PERF", # Perflint
    "PL",   # Pylint
    "UP",   # pyupgrade
]
ignore = ["PERF203", "PLC0415", "PLR0402", "PLW0603", "PLW0602"]

[tool.ruff.lint.mccabe]
max-complexity = 24

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.pylint]
allow-magic-value-types = ["bytes", "float", "int", "str"]
max-args = 23
max-branches = 23
max-returns = 13
max-statements = 102

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportUnnecessaryComparison = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportPrivateUsage = false
reportArgumentType = false
reportOptionalMemberAccess = false
reportReturnType = false
reportAttributeAccessIssue = false

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --tb=short"
timeout = 30
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["sandbox_runtime"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
